requirements.txt
fastapi==0.115.6
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
mysql-connector-python==8.0.29
pydantic==2.10.4
pydantic_core==2.27.2
typing_extensions>=4.12.2
bcrypt==4.1.2
starlette==0.37.0itsdangerous